chore: install circuit-ui
