+++ package.json
@@ -11,6 +11,14 @@
     "format": "prettier --write ."
   },
   "dependencies": {
+    "@emotion/core": "10.1.1",
+    "@emotion/styled": "^10.0.17",
+    "@sumup/circuit-ui": "^2.7.0",
+    "@sumup/collector": "^1.3.0",
+    "@sumup/design-tokens": "^2.2.0",
+    "@sumup/icons": "^1.7.0",
+    "@sumup/intl": "^1.2.0",
+    "emotion-theming": "^10.0.19",
     "firebase": "^8.6.8",
     "firebase-admin": "^9.10.0",
     "next": "11.0.1",
